stok işlemleri ile iligli kısım tamamlandı.

+++ ProductionApp.DTOs/StockDtos/StockAddDto.cs
@@ -8,7 +8,7 @@ public class StockAddDto : IBaseDto
     public string? Name { get; set; }
     public string? Description { get; set; }
     public decimal Price { get; set; }
-    public decimal VAT { get; set; }
+    public int VAT { get; set; }
     public DateTime CreatedDate { get; set; }
 
     public int GroupId { get; set; }
