+++ ProductionApp.Entities/Models/Stock.cs
@@ -9,7 +9,7 @@ public class Stock : IBaseEntity
     public string? Name { get; set; }
     public string? Description { get; set; }
     public decimal Price { get; set; }
-    public decimal VAT { get; set; }
+    public int VAT { get; set; }
     public DateTime CreatedDate { get; set; }
     public DateTime? ModifiedDate { get; set; }
